Add login form
Created HTML and CSS for a simple log in form. Users are required to input
a username and a password. There is also a link at the bottom that users
can click to create an account if they do not have one.

Once I hook up my django code with my javascript code, I will ensure that
users who have created an account will be able to sign in on their local
device.