Add sign-up form
Created a sign-up form that is similar to the log-in one. Basic design and
will allow a user to return to the log-in screen if they already have an
account.

Once I connect Django with my JavaScript, I will make sure that when users
create an account, it will be remembered on local devices.